requirements.txt
selenium>=4.15.0
openai>=1.3.0
tqdm>=4.66.0
requests>=2.31.0
flask-cors>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.2.0
supabase>=2.0.0
Pylette>=1.0.0
python-dotenv>=1.0.0 